src/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── MCP ───────────────────────────────────────────────────────────────────────
fastmcp>=2.3.0

# ── LangChain Ecosystem ──────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-neo4j>=0.3.0
langchain-ollama>=0.2.0

# ── LangGraph + SQLite Checkpointer ─────────────────────────────────────────
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=3.0.0
aiosqlite>=0.22.0

# ── Neo4j Driver ─────────────────────────────────────────────────────────────
neo4j>=5.19.0

# ── Config & Environment ─────────────────────────────────────────────────────
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── HTTP Client ──────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Utilities ────────────────────────────────────────────────────────────────
tenacity>=8.3.0

# ── Rate Limiting ────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── Structured Logging ───────────────────────────────────────────────────────
structlog>=24.0.0
python-json-logger>=2.0.0

# ── Observability ────────────────────────────────────────────────────────────
prometheus-fastapi-instrumentator>=7.0.0
